Logic for fetching and rendering portfolios moved from 'portfolios" to "homefeed"
